refactor: refactored upload image
