+++ web/src/components/rhf/file-upload/index.tsx
@@ -1,67 +1,22 @@
-import { scale } from "@cloudinary/url-gen/actions/resize"
-import { AdvancedImage } from "@cloudinary/react"
-import { Cloudinary, CloudinaryImage } from "@cloudinary/url-gen"
 import { Box, Button, IconButton, Stack, TextField } from "@mui/material"
-import { FunctionComponent, useEffect, useId, useState } from "react"
+import { CloudinaryImage } from "@cloudinary/url-gen"
+import { FunctionComponent, useState } from "react"
 import { Controller, useFormContext } from "react-hook-form"
 import RenderList from "@components/render-list"
-import useUploadImage from "@hooks/use-upload-image"
 import Render from "@components/render"
-import { byRadius } from "@cloudinary/url-gen/actions/roundCorners"
 import RenderIcon from "@components/render-icon"
 import icons from "@/icons"
+import { AdvancedImage } from "@cloudinary/react"
+import useImage from "@hooks/use-image"
 
 type Props = {
     label: string
     name: string
 }
 
-const genImageFromPublicID = (publicId: string): CloudinaryImage => {
-    const cld = new Cloudinary({ cloud: { cloudName: "commercium" } })
-    return cld
-        .image(publicId)
-        .resize(scale().width(100).height(100))
-        .roundCorners(byRadius(15))
-}
-
 const RHFImageUpload: FunctionComponent<Props> = ({ name, label }) => {
-    const { control, getValues, setValue } = useFormContext()
-    const mutation = useUploadImage()
-    const publicIdList: string[] = getValues("image")
-    const id = useId()
-
-    const [images, setImages] = useState<ImagePreview[]>([])
-    useEffect(() => {
-        setImages(
-            publicIdList.map((publicId, i) => {
-                return {
-                    id: `${id}${i}`,
-                    image: publicId ? genImageFromPublicID(publicId) : ""
-                }
-            })
-        )
-    }, [publicIdList])
-
-    const { status, data, variables } = mutation
-
-    useEffect(() => {
-        if (status === "success") {
-            setValue("image", [...getValues("image"), data.public_id])
-            setImages(prev => {
-                const temp = [...prev]
-                const imageIndex = temp.findIndex(
-                    ({ id }) => id === variables.id
-                )
-                const updatedImage = temp[imageIndex]
-                updatedImage.image = genImageFromPublicID(data.public_id)
-                temp.splice(imageIndex, 1, updatedImage)
-                return temp
-            })
-            if (!getValues("primary_image")) {
-                setValue("primary_image", data.secure_url)
-            }
-        }
-    }, [status, data, getValues, variables])
+    const { control } = useFormContext()
+    const { images, handleUpload } = useImage()
 
     return (
         <>
@@ -109,17 +64,7 @@ const RHFImageUpload: FunctionComponent<Props> = ({ name, label }) => {
                                     const { files } =
                                         e.target as HTMLInputElement
                                     if (files?.length) {
-                                        const itemID = `${id}${images.length}`
-                                        const payload = {
-                                            id: itemID,
-                                            file: files[0]
-                                        }
-
-                                        setImages([
-                                            ...images,
-                                            { id: itemID, image: "" }
-                                        ])
-                                        mutation.mutate(payload)
+                                        handleUpload(files)
                                     }
                                 }}
                                 label={label}
@@ -136,7 +81,6 @@ const RHFImageUpload: FunctionComponent<Props> = ({ name, label }) => {
 
 const ImageMiniPreview = ({ image }: { image: CloudinaryImage | string }) => {
     const [onHover, setOnHover] = useState(false)
-    console.log(image)
     return (
         <Box
             sx={{
